MaceStyleValidator/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
python-docx==1.1.2
vsdx==0.5.8
msal==1.31.1
requests==2.32.3